Feat: CSのリモコン番号をSID固定割当に (=3桁チャンネル番号)
CS110では運用規定上3桁チャンネル番号=SIDのため、BSのTR-B15固定割当に
倣い、CSはSIDそのものをリモコン番号として保存する。

- ChannelInfo.remote_control_key を u8→u16 に拡張 (CSのSIDは255超のため)。
  この構造体はワイヤエンコードされないためプロトコル互換に影響なし
- ワイヤ側 ClientChannelInfo は u8 のまま維持し、詰め替え時に255超は
  切り捨て値ではなく None を送る (session.rs / types.rs の2箇所)
- ChannelRecord (proxy) / Web API レスポンスも u16 に追従。
  CLI (recisdb) 側レコードは u8 のまま to_channel_info で widening

+++ recisdb-proxy/src/database/channel.rs
@@ -972,7 +972,7 @@ impl Database {
             raw_name: row.get("raw_name")?,
             channel_name: row.get("channel_name")?,
             physical_ch: row.get::<_, Option<i32>>("physical_ch")?.map(|v| v as u8),
-            remote_control_key: row.get::<_, Option<i32>>("remote_control_key")?.map(|v| v as u8),
+            remote_control_key: row.get::<_, Option<i32>>("remote_control_key")?.map(|v| v as u16),
             service_type: row.get::<_, Option<i32>>("service_type")?.map(|v| v as u8),
             network_name: row.get("network_name")?,
             bon_space: row.get::<_, Option<i32>>("bon_space")?.map(|v| v as u32),
